Trunk Sync - r5888 Special notes: MSVC_6_Release was corrected by hand, but none of the changes made to MSVC_6 were carried over.

+++ code/cutscene/cutscenes.cpp
@@ -653,8 +653,8 @@ void cutscenes_screen_do_frame()
 
 		Description_index = Selected_line;
 		Text_size = 0;
-		if ( Description_index < Cutscene_list.size( ) &&
-			 Cutscene_list[ Description_index ] < Cutscenes.size( ) ) {
+		if ( Description_index < (int)Cutscene_list.size( ) &&
+			 (int)Cutscene_list[ Description_index ] < (int)Cutscenes.size( ) ) {
 			src = Cutscenes[Cutscene_list[Description_index]].description;
 			if (src) {
 				Text_size = split_str(src, Cutscene_desc_coords[gr_screen.res][2], Text_line_size, Text_lines, Cutscene_max_text_lines[gr_screen.res]);
